pyproject.toml
[project]
name = "pybehave-selenium-test-framework"
version = "1.0.5"
authors = [
  { name="Argo DevOps", email="enquiry@argodevops.co.uk" },
]
description = "A powerful test automation framework designed to provide a comprehensive solution for testing user interfaces and APIs"
readme = "README.md"
license = {file = "LICENSE.txt"}
keywords = ["selenium", "behave"]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
  "allure-behave>=2.13.2",
  "allure-python-commons>=2.13.2",
  "art>=6.1",
  "attrs>=23.1.0",
  "azure-cli>=2.53.1",
  "azure-identity>=1.16.1",
  "azure-keyvault>=4.2.0",
  "azure-keyvault-secrets>=4.7.0",
  "behave>=1.2.6",
  "behavex>=2.0.1",
  "black>=24.3.0",
  "ddt>=1.6.0",
  "elementpath>=4.1.4",
  "Faker>=19.1.0",
  "iniconfig>=2.0.0",
  "Jinja2>=3.1.5",
  "packaging>=23.1",
  "panda>=0.3.1",
  "paramiko>=3.4.0",
  "parse>=1.19.1",
  "parse-type>=0.6.2",
  "pdfminer.six>=20240706",
  "pillow>=10.3.0", # not directly required, pinned by Snyk to avoid a vulnerability
  "pluggy>=1.2.0",
  "pre-commit>=3.3.3",
  "py>=1.11.0",
  "pyautogui>=0.9.54",
  "pygount>=1.2.4",
  "PyHamcrest>=2.1.0",
  "pylint>=2.17.4",
  "pyodbc>=4.0.39",
  "pyparsing>=3.1.0",
  "pytest>=7.4.0",
  "pyupgrade>=3.15.0",
  "requests>=2.32.0",
  "ruff>=0.2.1",
  "selenium>=4.15.2",
  "Selenium-Screenshot>=2.1.0",
  "setuptools>=65.5.1", # not directly required, pinned by Snyk to avoid a vulnerability
  "shortuuid>=1.0.11",
  "six>=1.16.0",
  "testresources>=2.0.1",
  "toml>=0.10.2",
  "typing-extensions>=4.7.1",
  "urllib3>=2.2.2",
  "wily>=1.24.2",
  "xmltodict>=0.13.0",
  "numpy>=1.22.2", # not directly required, pinned by Snyk to avoid a vulnerability
]

[project.urls]
Homepage = "https://github.com/argodevops/pytest-behave"
Issues = "https://github.com/argodevops/pytest-behave/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pybehave"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120

# Assume Python 3.10
target-version = "py310"
